fix Java runtime test regressions
