+++ tests/runtime/numbers/NumbersImpl.java
@@ -1,58 +1,10 @@
 package wit_numbers;
 
 public class NumbersImpl {
-    public static byte roundtripU8(byte a) {
-        return a;
-    }
-
-    public static byte roundtripS8(byte a) {
-        return a;
-    }
-
-    public static short roundtripU16(short a) {
-        return a;
-    }
-
-    public static short roundtripS16(short a) {
-        return a;
-    }
-
-    public static int roundtripU32(int a) {
-        return a;
-    }
-
-    public static int roundtripS32(int a) {
-        return a;
-    }
-
-    public static long roundtripU64(long a) {
-        return a;
-    }
-
-    public static long roundtripS64(long a) {
-        return a;
-    }
-
-    public static float roundtripFloat32(float a) {
-        return a;
-    }
-
-    public static double roundtripFloat64(double a) {
-        return a;
-    }
-
-    public static int roundtripChar(int a) {
-        return a;
-    }
-
-    private static int scalar = 0;
-
-    public static void setScalar(int a) {
-        scalar = a;
-    }
-
-    public static int getScalar() {
-        return scalar;
+    private static void expect(boolean v) {
+        if (!v) {
+            throw new AssertionError();
+        }
     }
 
     public static void testImports() {
@@ -108,9 +60,4 @@ public class NumbersImpl {
         expect(Imports.getScalar() == 4);
     }
 
-    private static void expect(boolean v) {
-        if (!v) {
-            throw new AssertionError();
-        }
-    }
 }
